updated dashboard component
